Backport 2c509a158fad63e69a8072fa4a7588eaacf37dc0
