+++ test/jdk/tools/jpackage/helpers/jdk/jpackage/test/TKit.java
@@ -40,10 +40,12 @@ import java.nio.file.WatchKey;
 import java.nio.file.WatchService;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.Date;
+import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -57,6 +59,7 @@ import java.util.function.Consumer;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
+import static java.util.stream.Collectors.toSet;
 import java.util.stream.Stream;
 import jdk.jpackage.test.Functional.ExceptionBox;
 import jdk.jpackage.test.Functional.ThrowingConsumer;
@@ -730,6 +733,94 @@ final public class TKit {
     public static void assertUnexpected(String msg) {
         currentTest.notifyAssert();
         error(concatMessages("Unexpected", msg));
+    }
+
+    public static DirectoryContentVerifier assertDirectoryContent(Path dir) {
+        return new DirectoryContentVerifier(dir);
+    }
+
+    public static final class DirectoryContentVerifier {
+        public DirectoryContentVerifier(Path baseDir) {
+            this(baseDir, ThrowingSupplier.toSupplier(() -> {
+                try (var files = Files.list(baseDir)) {
+                    return files.map(Path::getFileName).collect(toSet());
+                }
+            }).get());
+        }
+
+        public void match(Path ... expected) {
+            DirectoryContentVerifier.this.match(Set.of(expected));
+        }
+
+        public void match(Set<Path> expected) {
+            currentTest.notifyAssert();
+
+            var comm = Comm.compare(content, expected);
+            if (!comm.unique1().isEmpty() && !comm.unique2().isEmpty()) {
+                error(String.format(
+                        "assertDirectoryContentEquals(%s): Some expected %s. Unexpected %s. Missing %s",
+                        baseDir, format(comm.common()), format(comm.unique1()), format(comm.unique2())));
+            } else if (!comm.unique1().isEmpty()) {
+                error(String.format(
+                        "assertDirectoryContentEquals(%s): Expected %s. Unexpected %s",
+                        baseDir, format(comm.common()), format(comm.unique1())));
+            } else if (!comm.unique2().isEmpty()) {
+                error(String.format(
+                        "assertDirectoryContentEquals(%s): Some expected %s. Missing %s",
+                        baseDir, format(comm.common()), format(comm.unique2())));
+            } else {
+                traceAssert(String.format(
+                        "assertDirectoryContentEquals(%s): Expected %s",
+                        baseDir, format(expected)));
+            }
+        }
+
+        public void contains(Path ... expected) {
+            contains(Set.of(expected));
+        }
+
+        public void contains(Set<Path> expected) {
+            currentTest.notifyAssert();
+
+            var comm = Comm.compare(content, expected);
+            if (!comm.unique2().isEmpty()) {
+                error(String.format(
+                        "assertDirectoryContentContains(%s): Some expected %s. Missing %s",
+                        baseDir, format(comm.common()), format(comm.unique2())));
+            } else {
+                traceAssert(String.format(
+                        "assertDirectoryContentContains(%s): Expected %s",
+                        baseDir, format(expected)));
+            }
+        }
+
+        public DirectoryContentVerifier removeAll(Collection<Path> paths) {
+            Set<Path> newContent = new HashSet<>(content);
+            newContent.removeAll(paths);
+            return new DirectoryContentVerifier(baseDir, newContent);
+        }
+
+        public DirectoryContentVerifier removeAll(Path ... paths) {
+            return removeAll(List.of(paths));
+        }
+
+        public Set<Path> items() {
+            return content;
+        }
+
+        private DirectoryContentVerifier(Path baseDir, Set<Path> contents) {
+            this.baseDir = baseDir;
+            this.content = contents;
+        }
+
+        private static String format(Set<Path> paths) {
+            return Arrays.toString(
+                    paths.stream().sorted().map(Path::toString).toArray(
+                            String[]::new));
+        }
+
+        private final Path baseDir;
+        private final Set<Path> content;
     }
 
     public static void assertStringListEquals(List<String> expected,
