+++ test/jdk/tools/jpackage/share/DotInNameTest.java
@@ -21,7 +21,6 @@
  * questions.
  */
 
-package jdk.jpackage.tests;
 
 import jdk.jpackage.test.JPackageCommand;
 import jdk.jpackage.test.HelloApp;
@@ -37,7 +36,7 @@ import jdk.jpackage.test.Annotations.Test;
  * @build jdk.jpackage.test.*
  * @compile DotInNameTest.java
  * @run main/othervm/timeout=360 -Xmx512m jdk.jpackage.test.Main
- *  --jpt-run=jdk.jpackage.tests.DotInNameTest
+ *  --jpt-run=DotInNameTest
  *  --jpt-before-run=jdk.jpackage.test.JPackageCommand.useToolProviderByDefault
  */
