+++ test/jdk/tools/jpackage/share/ModulePathTest.java
@@ -21,7 +21,6 @@
  * questions.
  */
 
-package jdk.jpackage.tests;
 
 import java.io.File;
 import java.io.IOException;
@@ -48,7 +47,7 @@ import jdk.jpackage.test.Annotations.Test;
  * @build jdk.jpackage.test.*
  * @compile ModulePathTest.java
  * @run main/othervm/timeout=360 -Xmx512m jdk.jpackage.test.Main
- *  --jpt-run=jdk.jpackage.tests.ModulePathTest
+ *  --jpt-run=ModulePathTest
  */
 
 public final class ModulePathTest {
