+++ test/jdk/tools/jpackage/windows/WinL10nTest.java
@@ -22,7 +22,6 @@
  */
 
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import jdk.jpackage.test.TKit;
 import jdk.jpackage.test.PackageTest;
@@ -35,8 +34,6 @@ import java.util.List;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 import jdk.jpackage.test.Executor;
-
-import static jdk.jpackage.test.WindowsHelper.getTempDirectory;
 
 /*
  * @test
@@ -163,9 +160,8 @@ public class WinL10nTest {
             }
 
             // Preserve config dir to check the set of copied l10n files.
-            Path tempDir = getTempDirectory(cmd, tempRoot);
-            Files.createDirectories(tempDir.getParent());
-            cmd.addArguments("--temp", tempDir.toString());
+            Path tempDir = tempRoot.resolve(cmd.packageType().name());
+            cmd.addArguments("--temp", tempDir);
         })
         .addBundleVerifier((cmd, result) -> {
             if (expectedCulture != null) {
@@ -186,7 +182,7 @@ public class WinL10nTest {
                                     getLightCommandLine(result));
                         }
                     }
-                    Path tempDir = getTempDirectory(cmd, tempRoot).toAbsolutePath();
+                    var tempDir = Path.of(cmd.getArgumentValue("--temp")).toAbsolutePath();
                     for (var v : createDefaultL10nFilesLocVerifiers(tempDir)) {
                         v.apply(getLightCommandLine(result));
                     }
